Make websockets work enough to demonstrate
This commit captures a full days work trying to make changes to make it so that the joining of a game or hosting/creating a game is:

* allowing the host to be added to the list of players
* updating for all clients when a new player joins
* using signals and things related to signals to be a bit more modern than using subscription and a bunch of rxjs things (though this does seem to ignore errors a bit more...)

Perhaps the most helpful bit could be the use of a snapshot of the params to hold onto the id of the game in the game component... that was not obvious to me, and I kept struggling to find the id and it's just RIGHT THERE so that was frustrating... it's so visible and present, but I didn't have a tidy way of accessing it.

I think the communication for web sockets may be able to be fixed a bit more to only broadcast to things that are likely to care, but maybe there will be other use cases I'm not thinking about that will want to tell about changes in the database that happen and may impact lots of components... so maybe this will not be a good thing to worry about.

I most regret that I don't have any tests for this. That will be my next big thing to look at.