Change order of steps in join process
There were two web socket messages on the connection that originated the add player action, and I think it was because I was first navigating to the new page and then sending the web socket message. What I should have done (probably) is send the web socket message and then navigate to the new page. This didn't prevent the handling from trying to update even before the game was instantiated on the new page, but it at least seemed helpful.