Streamline use of web socket message to update game signal
I think I was using signals in a somewhat incorrect or discouraged way. My research had revealed that I should probably avoid directly changing the object or its properties. So, I think my use of push on the array of players was considered a bit dangerous/incorrect. I know it felt a bit strange to have the client update something that was not coming from the database. I have helped myself come to terms with the way this works by (trying out and removing) having the version of the playerName that is put into the player array in the component view of the data include some other text... once the component changes for itself (instead of via web socket messages), the extra text goes away and it's just the actual player name that is in the array. I loved this because it showed me that I could do something like say "this player was added since you last updated anything directly from this component in this client" -- it allows you to see the data that is new, it updates to what the database has when it next checks in directly, and it doesn't add both the extra text version and the regular version anywhere. When I accidentally added a name locally and also remotely, the local one didn't seem to update correctly after that (the array still included the one I added directly to the array).