Add "boom!" for fun and to see difference between ws and database
This addition makes it easier to notice when a new player is added and the update came from a web socket message. When the page is refreshed on a client or something changes on a specific client that causes a database kind of action, the "boom" part goes away since it was only local (and was never sent to the database).