Move/add web socket for hosting into host
If we want to keep track of the users using web sockets, we need to set up that web socket when we have access for the name, and I'm not sure how to access it in the game component (it seems like we don't know when the message will be sent in relation to other things). Probably need to also use localStorage and sessionStorage to help with that.