Avoid memory leaks by using `takeUntilDestroyed`
The shift away from using lifecycle hooks may have left us a little vulnerable to memory leaks. From what I understand, one thing we can do that is relatively new is to use takeUntilDestroyed, which simplifies the unsubscribe process. I'm not totally sure it's necessary, but it doesn't seem to break anything, so I'm giving it a try.