Change from node to bundler
I'm unsure about this change. In the past, We had some trouble with updating the builder, and I'm not sure how this relates to that. I went ahead and updated it in the way that the `ng update` did automatically (after I manually changed it to node when the misbehaving user test failed). Keeping this a separate commit since I don't feel confident about it.