Update Angular - automatic stuff
I *think* this part has to do with how things will be named when you generate them using `ng generate`. My understanding is that some of the naming schemes have changed -- where one used to use the type of what something was right in the name, that happens less often now. I'm not sure: https://angular.dev/style-guide#match-file-names-to-the-typescript-identifier-within